Crate remote_externalities

Remote Externalities

An equivalent of sp_io::TestExternalities that can load its state from a remote substrate based chain, or a local state snapshot file.

Modules

rpc_api

WS RPC API for one off RPC calls to a substrate node.

Structs

Builder

Builder for remote-externalities.

OfflineConfig

Configuration of the offline execution.

OnlineConfig

Configuration of the online execution.

SnapshotConfig

Configuration of the state snapshot.

Transport

Description of the transport protocol (for online execution).

Enums

Mode

The execution mode.

Traits

RpcApiClient

Client implementation for the RpcApi RPC API.

Type Definitions

TestExternalities

Type alias for Externalities implementation used in tests.