Crate remote_externalities

Remote Externalities

An equivalent of sp_io::TestExternalities that can load its state from a remote substrate based chain, or a local state snapshot file.

Structs

Builder

Builder for remote-externalities.

OfflineConfig

Configuration of the offline execution.

OnlineConfig

Configuration of the online execution.

SnapshotConfig

Configuration of the state snapshot.

Enums

Mode

The execution mode.

Transport

Description of the transport protocol (for online execution).

Type Definitions

TestExternalities

Type alias for Externalities implementation used in tests.